refactor: move GroupedEvaluator class to utility functions in utils/group_eval.py
